[prerelease] Fix hook exports
