+++ webpack/published-js-components.js
@@ -1,8 +1,9 @@
 const publishedJSComponents = {
-  ResponsiveList: "components/ResponsiveList/ResponsiveList.jsx",
-  MultiStepIndicator: "components/MultiStepIndicator/MultiStepIndicator.jsx",
-  Dropdown: "components/Dropdown/Dropdown.jsx",
-  GridKeyboardNavigationContext: "components/GridKeyboardNavigationContext/GridKeyboardNavigationContext.js",
+  ResponsiveList: "components/ResponsiveList/ResponsiveList",
+  MultiStepIndicator: "components/MultiStepIndicator/MultiStepIndicator",
+  Dropdown: "components/Dropdown/Dropdown",
+  GridKeyboardNavigationContext: "components/GridKeyboardNavigationContext/GridKeyboardNavigationContext",
+  useGridKeyboardNavigationContext: "hooks/useGridKeyboardNavigationContext/useGridKeyboardNavigationContext",
   allIcons: "components/Icon/Icons"
 };
 // eslint-disable-next-line no-undef
