+++ webpack/published-ts-components.js
@@ -9,7 +9,7 @@ const publishedTSComponents = {
   IconButton: "components/IconButton/IconButton",
   Link: "components/Link/Link",
   Combobox: "components/Combobox/Combobox",
-  ColorUtils: "utils/colors-utils.ts",
+  ColorUtils: "utils/colors-utils",
   Counter: "components/Counter/Counter",
   Divider: "components/Divider/Divider",
   FormattedNumber: "components/FormattedNumber/FormattedNumber",
@@ -70,8 +70,8 @@ const publishedTSComponents = {
   VirtualizedList: "components/VirtualizedList/VirtualizedList",
   VirtualizedGrid: "components/VirtualizedGrid/VirtualizedGrid",
   LinearProgressBar: "components/ProgressBars/LinearProgressBar/LinearProgressBar",
-  Dialog: "components/Dialog/Dialog.tsx",
-  MenuButton: "components/MenuButton/MenuButton.tsx",
+  Dialog: "components/Dialog/Dialog",
+  MenuButton: "components/MenuButton/MenuButton",
   Tipseen: "components/Tipseen/Tipseen",
   TipseenContent: "components/Tipseen/TipseenContent",
   TipseenWizard: "components/Tipseen/TipseenWizard",
@@ -85,18 +85,18 @@ const publishedTSComponents = {
   MenuGridItem: "components/Menu/MenuGridItem/MenuGridItem",
   // Don't remove next line
   // plop_marker:published-hooks
-  useKeyEvent: "hooks/useKeyEvent/index.ts",
-  useEventListener: "hooks/useEventListener/index.ts",
-  useDebounceEvent: "hooks/useDebounceEvent/index.ts",
-  useClickOutside: "hooks/useClickOutside/index.ts",
-  useResizeObserver: "hooks/useResizeObserver.ts",
-  useAfterFirstRender: "hooks/useAfterFirstRender/index.ts",
-  useTimeout: "hooks/useTimeout/index.ts",
-  usePrevious: "hooks/usePrevious/index.ts",
-  useSetFocus: "hooks/useSetFocus/index.ts",
-  useMergeRefs: "hooks/useMergeRefs.ts",
-  useIsMouseOver: "hooks/useIsMouseOver.ts",
-  useGridKeyboardNavigation: "hooks/useGridKeyboardNavigation/useGridKeyboardNavigation.ts",
+  useKeyEvent: "hooks/useKeyEvent/index",
+  useEventListener: "hooks/useEventListener/index",
+  useDebounceEvent: "hooks/useDebounceEvent/index",
+  useClickOutside: "hooks/useClickOutside/index",
+  useResizeObserver: "hooks/useResizeObserver",
+  useAfterFirstRender: "hooks/useAfterFirstRender/index",
+  useTimeout: "hooks/useTimeout/index",
+  usePrevious: "hooks/usePrevious/index",
+  useSetFocus: "hooks/useSetFocus/index",
+  useMergeRefs: "hooks/useMergeRefs",
+  useIsMouseOver: "hooks/useIsMouseOver",
+  useGridKeyboardNavigation: "hooks/useGridKeyboardNavigation/useGridKeyboardNavigation",
   useIsOverflowing: "hooks/useIsOverflowing/useIsOverflowing",
   useMediaQuery: "hooks/useMediaQuery",
   useVibeMediaQuery: "hooks/useVibeMediaQuery",
